safety-service/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

grpcio~=1.80.0
transformers~=5.5.0
validators~=0.35.0
torch==2.11.0